lambda/requirements.txt
ccxt
boto3
pandas
python-dotenv
